processor: fix global resolved worker deadlock
In globalResolvedWorker, the error group waits for input forward function returns.
The input forward function won't exit until the `outputTxn` is closed.
The outputTxn channel is closed in a defer function, which only runs after the `inputTxn` is closed

We have two ways to let input forward function returns, one way is add a context to
this function, the other way is close the `inputTxn` when puller exits.