cdc: Remove context.Canceled checks
The functions just simply return whatever errors they come across, let
the callers decide what to do with the errors.
In our case, we are calling these functions with an error group, and the
only situation these functions might get a context.Canceled is when we
call the cancel function of their common context, in which case we don't
care if the context.Cancled returned by one function causes the others
to be canceled by the error group.