*: support suspend jobs, part 1 (#79)
part job of #60
This PR supports the suspending of jobs and tasks, but it hasn't considered the consistency after fail-over.

The "suspend" is implemented in an async way, because suspending might happen when rescheduling or fail recovering.
After the user sends suspend request, we mark the "target status" for the target tasks. A goroutine will check the status regularly, and will try to suspend it util success. There are some work to do furture.

* The commands should be serializable. If several commands are sent in the same time, we should order them.
* The command have to be persisted and to be continued when node crashes.